stable_diffusion_py/mac_requirements.txt
pynecone==0.1.33a2
diffusers
pillow
transformers